requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
apscheduler==3.10.4
alembic==1.12.1
pytest==7.4.3